requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.11.0
click==8.1.7
crispy-bootstrap5==0.7
Django==4.2.6
django-crispy-forms==2.0
flake8==5.0.4
flake8-html==0.4.3
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==4.1.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.17.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
